initial styling and idealization for calendar page.
